fix tests (WIP)
